Motor_Control_485_QT/requirements.txt
PyQt5>=5.15.0
pyserial>=3.5
openpyxl>=3.1.0
